pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Анализатор страниц"
readme = "README.md"
authors = [{name = "pavel", email = "paveldidko@gmail.com"},]

dependencies = [
  "flask>=3.0.3",
  "gunicorn>=23.0.0",
  "psycopg2-binary>=2.9.9",
  "python-dotenv>=0.21.0",
  "requests>=2.32.4",
  "validators>=0.35.0",
]

[tool.uv]
dev-dependencies = [
  "flake8>=7.1.1"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
